EntityDeleteConfirmModalDisplay: move to @labkey/components/entities
